Support DCS 2.7 cloud presets.
In 2.7 cloud presets can be selected, which will:

* Override cloud thickness, density, and precipitation
* Bound the cloud base
* Set the "preset" property in the mission file

This patch adds an import script to load the names and cloud base
boundaries from the config file in the DCS install. The script requires
lupa because the parser in dcs.lua is not capable of parsing complex lua
files (I think it's because there are function calls, but I didn't dig
much).

This also removes the `Weather.fog_density` property, since DCS no
longer supports that.

Fixes https://github.com/pydcs/dcs/issues/118